Added 'sierpinski' as a shape
This may want to be changed to a fractal later, but for now it works.
There seemed to be some seg fault issues when I was testing, and I'm not
actually sure how I resolved them, but it's worked the last few times
I've tried it so I'm committing deal with it.