Text editor is more efficient/a bit less buggy
I defined an EMPTY_CELL string; if this is passed to the text editor
grid, it means there is nothing in there—not even “ “ (a space). This
is leading up to moving text around when entering letters, pressing
space, enter/tab etc.

It also means that if the EMPTY_CELL character is in a grid square, we
don’t have to bother rendering text to it—which has given the editor a
performance jump. Still need to make it more cached though.