Right, read the description:
So I’ve tried fixing the mac flickering with no result: Linux people
are going to have to feed back on what’s going on. (Side note, whilst
I’m in NL I’m going to have a go on a friend’s computer to test things
out in Linux).

On my Mac, there is reduced flickering on the canvas/challenge modes,
but when you press generate you get a flickering fractal (which is
rubbish).

The text editor works except for a few bugs (backspacing when there
isn’t very much in the text editor seems to be a big issue).

I’ve also given the menu screen a lick of paint: same treatment will be
applied to the interfaces this week.